+++ ref-exchange/src/utils.rs
@@ -9,6 +9,10 @@ use uint::construct_uint;
 
 /// Attach no deposit.
 pub const NO_DEPOSIT: u128 = 0;
+
+/// 10T gas for basic operation
+pub const GAS_FOR_BASIC_OP: Gas = 10_000_000_000_000;
+
 /// hotfix_insuffient_gas_for_mft_resolve_transfer.
 pub const GAS_FOR_RESOLVE_TRANSFER: Gas = 20_000_000_000_000;